Move to feature, add upload date, move few things to new line
